+++ spine-csharp/src/SkeletonData.cs
@@ -43,6 +43,7 @@ namespace Spine {
 		internal ExposedList<IkConstraintData> ikConstraints = new ExposedList<IkConstraintData>();
 		internal ExposedList<TransformConstraintData> transformConstraints = new ExposedList<TransformConstraintData>();
 		internal ExposedList<PathConstraintData> pathConstraints = new ExposedList<PathConstraintData>();
+		internal ExposedList<SpringConstraintData> springConstraints = new ExposedList<SpringConstraintData>();
 		internal float x, y, width, height;
 		internal string version, hash;
 
@@ -95,7 +96,7 @@ namespace Spine {
 		/// <summary>The dopesheet FPS in Spine, or zero if nonessential data was not exported.</summary>
 		public float Fps { get { return fps; } set { fps = value; } }
 
-		// --- Bones.
+		// --- Bones
 
 		/// <summary>
 		/// Finds a bone by comparing each bone's name.
@@ -111,7 +112,7 @@ namespace Spine {
 			return null;
 		}
 
-		// --- Slots.
+		// --- Slots
 
 		/// <returns>May be null.</returns>
 		public SlotData FindSlot (string slotName) {
@@ -124,7 +125,7 @@ namespace Spine {
 			return null;
 		}
 
-		// --- Skins.
+		// --- Skins
 
 		/// <returns>May be null.</returns>
 		public Skin FindSkin (string skinName) {
@@ -134,7 +135,7 @@ namespace Spine {
 			return null;
 		}
 
-		// --- Events.
+		// --- Events
 
 		/// <returns>May be null.</returns>
 		public EventData FindEvent (string eventDataName) {
@@ -144,7 +145,7 @@ namespace Spine {
 			return null;
 		}
 
-		// --- Animations.
+		// --- Animations
 
 		/// <returns>May be null.</returns>
 		public Animation FindAnimation (string animationName) {
@@ -157,7 +158,7 @@ namespace Spine {
 			return null;
 		}
 
-		// --- IK constraints.
+		// --- IK constraints
 
 		/// <returns>May be null.</returns>
 		public IkConstraintData FindIkConstraint (string constraintName) {
@@ -170,7 +171,7 @@ namespace Spine {
 			return null;
 		}
 
-		// --- Transform constraints.
+		// --- Transform constraints
 
 		/// <returns>May be null.</returns>
 		public TransformConstraintData FindTransformConstraint (string constraintName) {
@@ -183,14 +184,35 @@ namespace Spine {
 			return null;
 		}
 
-		// --- Path constraints.
+		// --- Path constraints
 
+		/// <summary>
+		/// Finds a path constraint by comparing each path constraint's name. It is more efficient to cache the results of this method
+		/// than to call it multiple times.
+		/// </summary>
 		/// <returns>May be null.</returns>
 		public PathConstraintData FindPathConstraint (string constraintName) {
 			if (constraintName == null) throw new ArgumentNullException("constraintName", "constraintName cannot be null.");
 			var pathConstraints = this.pathConstraints.Items;
 			for (int i = 0, n = this.pathConstraints.Count; i < n; i++) {
 				PathConstraintData constraint = pathConstraints[i];
+				if (constraint.name.Equals(constraintName)) return constraint;
+			}
+			return null;
+		}
+
+		// --- Spring constraints
+
+		/// <summary>
+		/// Finds a spring constraint by comparing each spring constraint's name. It is more efficient to cache the results of this
+		/// method than to call it multiple times.
+		/// </summary>
+		/// <returns>May be null.</returns>
+		public SpringConstraintData FindSpringConstraint (String constraintName) {
+			if (constraintName == null) throw new ArgumentNullException("constraintName", "constraintName cannot be null.");
+			Object[] springConstraints = this.springConstraints.Items;
+			for (int i = 0, n = this.springConstraints.Count; i < n; i++) {
+				SpringConstraintData constraint = (SpringConstraintData)springConstraints[i];
 				if (constraint.name.Equals(constraintName)) return constraint;
 			}
 			return null;