+++ spine-unity/Assets/Spine/Runtime/spine-unity/Utility/AttachmentRegionExtensions.cs
@@ -33,58 +33,6 @@ using UnityEngine;
 
 namespace Spine.Unity.AttachmentTools {
 	public static class AttachmentRegionExtensions {
-		#region SetRegion
-		/// <summary>
-		/// Tries to set the region (image) of a renderable attachment. If the attachment is not renderable, nothing is applied.</summary>
-		public static void SetRegion (this Attachment attachment, AtlasRegion region, bool updateOffset = true) {
-			var regionAttachment = attachment as RegionAttachment;
-			if (regionAttachment != null)
-				regionAttachment.SetRegion(region, updateOffset);
-
-			var meshAttachment = attachment as MeshAttachment;
-			if (meshAttachment != null)
-				meshAttachment.SetRegion(region, updateOffset);
-		}
-
-		/// <summary>Sets the region (image) of a RegionAttachment</summary>
-		public static void SetRegion (this RegionAttachment attachment, AtlasRegion region, bool updateOffset = true) {
-			if (region == null) throw new System.ArgumentNullException("region");
-
-			// (AtlasAttachmentLoader.cs)
-			attachment.RendererObject = region;
-			attachment.SetUVs(region.u, region.v, region.u2, region.v2, region.degrees);
-			attachment.RegionOffsetX = region.offsetX;
-			attachment.RegionOffsetY = region.offsetY;
-			attachment.RegionWidth = region.width;
-			attachment.RegionHeight = region.height;
-			attachment.RegionOriginalWidth = region.originalWidth;
-			attachment.RegionOriginalHeight = region.originalHeight;
-
-			if (updateOffset) attachment.UpdateOffset();
-		}
-
-		/// <summary>Sets the region (image) of a MeshAttachment</summary>
-		public static void SetRegion (this MeshAttachment attachment, AtlasRegion region, bool updateUVs = true) {
-			if (region == null) throw new System.ArgumentNullException("region");
-
-			// (AtlasAttachmentLoader.cs)
-			attachment.RendererObject = region;
-			attachment.RegionU = region.u;
-			attachment.RegionV = region.v;
-			attachment.RegionU2 = region.u2;
-			attachment.RegionV2 = region.v2;
-			attachment.RegionDegrees = region.degrees;
-			attachment.RegionOffsetX = region.offsetX;
-			attachment.RegionOffsetY = region.offsetY;
-			attachment.RegionWidth = region.width;
-			attachment.RegionHeight = region.height;
-			attachment.RegionOriginalWidth = region.originalWidth;
-			attachment.RegionOriginalHeight = region.originalHeight;
-
-			if (updateUVs) attachment.UpdateUVs();
-		}
-		#endregion
-
 		#region Runtime RegionAttachments
 		/// <summary>
 		/// Creates a RegionAttachment based on a sprite. This method creates a real, usable AtlasRegion. That AtlasRegion uses a new AtlasPage with the Material provided./// </summary>
@@ -128,15 +76,7 @@ namespace Spine.Unity.AttachmentTools {
 			// (AtlasAttachmentLoader.cs)
 			var attachment = new RegionAttachment(attachmentName);
 
-			attachment.RendererObject = region;
-			attachment.SetUVs(region.u, region.v, region.u2, region.v2, region.degrees);
-			attachment.RegionOffsetX = region.offsetX;
-			attachment.RegionOffsetY = region.offsetY;
-			attachment.RegionWidth = region.width;
-			attachment.RegionHeight = region.height;
-			attachment.RegionOriginalWidth = region.originalWidth;
-			attachment.RegionOriginalHeight = region.originalHeight;
-
+			attachment.Region = region;
 			attachment.Path = region.name;
 			attachment.ScaleX = 1;
 			attachment.ScaleY = 1;
@@ -148,11 +88,15 @@ namespace Spine.Unity.AttachmentTools {
 			attachment.A = 1;
 
 			// pass OriginalWidth and OriginalHeight because UpdateOffset uses it in its calculation.
-			attachment.Width = attachment.RegionOriginalWidth * scale;
-			attachment.Height = attachment.RegionOriginalHeight * scale;
+			var textreRegion = attachment.Region;
+			var atlasRegion = textreRegion as AtlasRegion;
+			float originalWidth = atlasRegion != null ? atlasRegion.originalWidth : textreRegion.width;
+			float originalHeight = atlasRegion != null ? atlasRegion.originalHeight : textreRegion.height;
+			attachment.Width = originalWidth * scale;
+			attachment.Height = originalHeight * scale;
 
 			attachment.SetColor(Color.white);
-			attachment.UpdateOffset();
+			attachment.UpdateRegion();
 			return attachment;
 		}